Solving the issue with th postprocessing.
Problem:
In the new version modified for improving the regional trades, as not
all the carriers will be activated for the trade, some modification was
needed in the postprocessing phase.

Solution:
In the previous version, the results datafraes for the trade variables
were taking all the carriers in the model as columns while in the new
version, only user specificed regions between two nodes should be taken
as column.

Modules under change:
1. postprocess.py -> modifying the columns for the trade vars
2. main.py -> to pass the regional matches for traded carriers