Adding coordination to the nodes
problem:  In a multi-node model, the losses and unit costs of the inter-regional connections are mostly defined per unit of distance between two nodes that can highly affect the optimization.

Solution:
In this new version of the model, the user has the option to add the Lat and Long of each region in the global.xlsx file. The model will calculate the distance among each per of regions based on the Vincenty's inverse method inspired by calliope and the calculate the related variable and parameters based on the calculated or reported distance.

Modules under change:

1. utility.py -> adding the Vincenty's function.

2. constants.py -> adding the new parameter sheet name: line_length

3. StrData.py -> read the coordination sets and line's length parameter

4. Build.py -> Modify the trade efficiency and cost equations