allowing the model to chose among multiple size for each connection
Users can give the model the options of sizes for the inter-regional connections in the "global.xlsx" file. All the inter-regional parameters are defined for each given size. the "line_newcapacity" variable is a boolean variable and is defined for each size type, year, and traded carrier among the regions:

- Build.py:
1) modifying the line new capacity variable to a boolean type
2) Adding the "line_integer_size" constraint to make sure that the model will install only one size for each connection in each modelling year of the horizon
3) Modifying the cost functions based on the new type of variables

- StrData.py:
1) Reading the size options from set files
2) Modifying the parameter sheets

- constants.py: modifying the parameter sheets

- utility.py: Modifying the accumulated new capacity function for the connections